feat(api): add status to /mon/metrics
```
status{component="Global/Status",instance="cdsinstance",status="AL"} 1
status{component="Global/Version",instance="cdsinstance",status="OK"} 1
status{component="Global/api",instance="cdsinstance",status="AL"} 3
status{component="Global/dbmigrate",instance="cdsinstance",status="OK"} 0
status{component="Global/doc",instance="cdsinstance",status="WARN"} 2
status{component="Global/elasticsearch",instance="cdsinstance",status="OK"} 0
status{component="Global/hatchery",instance="cdsinstance",status="AL"} 0
status{component="Global/hooks",instance="cdsinstance",status="OK"} 1
status{component="Global/repositories",instance="cdsinstance",status="OK"} 0
status{component="Global/vcs",instance="cdsinstance",status="AL"} 1
status{component="api_0/CDSName",instance="cdsinstance",status="OK"} 1
status{component="api_0/Cache",instance="cdsinstance",status="OK"} 1
status{component="api_0/Database",instance="cdsinstance",status="OK"} 1
status{component="api_0/Event",instance="cdsinstance",status="OK"} 1
status{component="api_0/Hostname",instance="cdsinstance",status="OK"} 1
status{component="api_0/Internal Events Queue",instance="cdsinstance",status="OK"} 0
status{component="api_0/Nb of Panics",instance="cdsinstance",status="OK"} 0
status{component="api_0/Object-Store",instance="cdsinstance",status="OK"} 1
status{component="api_0/SMTP",instance="cdsinstance",status="AL"} 1
status{component="api_0/Scheduler",instance="cdsinstance",status="WARN"} 1
status{component="api_0/Sessions-Store",instance="cdsinstance",status="OK"} 1
status{component="api_0/Time",instance="cdsinstance",status="OK"} 1
status{component="api_0/Uptime",instance="cdsinstance",status="OK"} 1
status{component="api_0/Version",instance="cdsinstance",status="OK"} 1
status{component="api_0/Worker Model Errors",instance="cdsinstance",status="OK"} 0
status{component="api_1/CDSName",instance="cdsinstance",status="OK"} 1
status{component="api_1/Cache",instance="cdsinstance",status="OK"} 1
status{component="api_1/Database",instance="cdsinstance",status="OK"} 1
status{component="api_1/Event",instance="cdsinstance",status="OK"} 1
status{component="api_1/Hostname",instance="cdsinstance",status="OK"} 1
status{component="api_1/Internal Events Queue",instance="cdsinstance",status="OK"} 0
status{component="api_1/Nb of Panics",instance="cdsinstance",status="OK"} 0
status{component="api_1/Object-Store",instance="cdsinstance",status="OK"} 1
status{component="api_1/SMTP",instance="cdsinstance",status="AL"} 1
status{component="api_1/Scheduler",instance="cdsinstance",status="OK"} 1
status{component="api_1/Sessions-Store",instance="cdsinstance",status="OK"} 1
status{component="api_1/Time",instance="cdsinstance",status="OK"} 1
status{component="api_1/Uptime",instance="cdsinstance",status="OK"} 1
status{component="api_1/Version",instance="cdsinstance",status="OK"} 1
status{component="api_1/Worker Model Errors",instance="cdsinstance",status="OK"} 0
status{component="api_2/CDSName",instance="cdsinstance",status="OK"} 1
status{component="api_2/Cache",instance="cdsinstance",status="OK"} 1
status{component="api_2/Database",instance="cdsinstance",status="OK"} 1
status{component="api_2/Event",instance="cdsinstance",status="OK"} 1
status{component="api_2/Hostname",instance="cdsinstance",status="OK"} 1
status{component="api_2/Internal Events Queue",instance="cdsinstance",status="OK"} 0
status{component="api_2/Nb of Panics",instance="cdsinstance",status="OK"} 0
status{component="api_2/Object-Store",instance="cdsinstance",status="OK"} 1
status{component="api_2/SMTP",instance="cdsinstance",status="AL"} 1
status{component="api_2/Scheduler",instance="cdsinstance",status="OK"} 1
status{component="api_2/Sessions-Store",instance="cdsinstance",status="OK"} 1
status{component="api_2/Time",instance="cdsinstance",status="OK"} 1
status{component="api_2/Uptime",instance="cdsinstance",status="OK"} 1
status{component="api_2/Version",instance="cdsinstance",status="OK"} 1
status{component="api_2/Worker Model Errors",instance="cdsinstance",status="OK"} 0
status{component="hooksLocal/Time",instance="cdsinstance",status="OK"} 1
status{component="hooksLocal/Uptime",instance="cdsinstance",status="OK"} 1
status{component="hooksLocal/Version",instance="cdsinstance",status="OK"} 1
status{component="sample-service/sample-service/sample-service",instance="cdsinstance",status="WARN"} 1
status{component="sample-service2/sample-service2/sample-service2",instance="cdsinstance",status="WARN"} 1
status{component="vcsLocal/Github-RateLimit",instance="cdsinstance",status="AL"} 0
status{component="vcsLocal/Github-RateLimitRemaining",instance="cdsinstance",status="OK"} 5000
status{component="vcsLocal/Github-RateLimitReset",instance="cdsinstance",status="AL"} 1
status{component="vcsLocal/Time",instance="cdsinstance",status="OK"} 1
status{component="vcsLocal/Uptime",instance="cdsinstance",status="OK"} 1
status{component="vcsLocal/Version",instance="cdsinstance",status="OK"} 1
```
