+++ engine/api/status.go
@@ -3,13 +3,11 @@ package api
 import (
 	"bytes"
 	"context"
-	"fmt"
 	"net/http"
-	"sort"
-	"strings"
 
 	"github.com/ovh/cds/engine/api/event"
 	"github.com/ovh/cds/engine/api/mail"
+	"github.com/ovh/cds/engine/api/metrics"
 	"github.com/ovh/cds/engine/api/objectstore"
 	"github.com/ovh/cds/engine/api/repositoriesmanager"
 	"github.com/ovh/cds/engine/api/scheduler"
@@ -63,119 +61,9 @@ func (api *API) statusHandler() service.Handler {
 			return err
 		}
 
-		mStatus := api.computeGlobalStatus(srvs)
+		mStatus := metrics.ComputeGlobalStatus(srvs)
 		return service.WriteJSON(w, mStatus, status)
 	}
-}
-
-type computeGlobalNumbers struct {
-	nbSrv       int
-	nbOK        int
-	nbAlerts    int
-	nbWarn      int
-	minInstance int
-}
-
-func (api *API) computeGlobalStatus(srvs []sdk.Service) sdk.MonitoringStatus {
-	mStatus := sdk.MonitoringStatus{}
-
-	var version string
-	versionOk := true
-	linesGlobal := []sdk.MonitoringStatusLine{}
-
-	resume := map[string]computeGlobalNumbers{
-		services.TypeAPI:           {minInstance: api.Config.Status.API.MinInstance},
-		services.TypeRepositories:  {minInstance: api.Config.Status.Repositories.MinInstance},
-		services.TypeVCS:           {minInstance: api.Config.Status.VCS.MinInstance},
-		services.TypeHooks:         {minInstance: api.Config.Status.Hooks.MinInstance},
-		services.TypeHatchery:      {minInstance: api.Config.Status.Hatchery.MinInstance},
-		services.TypeDBMigrate:     {minInstance: api.Config.Status.DBMigrate.MinInstance},
-		services.TypeElasticsearch: {minInstance: api.Config.Status.ElasticSearch.MinInstance},
-	}
-	var nbg computeGlobalNumbers
-	for _, s := range srvs {
-		var nbOK, nbWarn, nbAlert int
-		for i := range s.MonitoringStatus.Lines {
-			l := s.MonitoringStatus.Lines[i]
-			mStatus.Lines = append(mStatus.Lines, l)
-
-			switch l.Status {
-			case sdk.MonitoringStatusOK:
-				nbOK++
-			case sdk.MonitoringStatusWarn:
-				nbWarn++
-			default:
-				nbAlert++
-			}
-
-			// services should have same version
-			if strings.Contains(l.Component, "Version") {
-				if version == "" {
-					version = l.Value
-				} else if version != l.Value && versionOk {
-					versionOk = false
-					linesGlobal = append(linesGlobal, sdk.MonitoringStatusLine{
-						Status:    sdk.MonitoringStatusWarn,
-						Component: "Global/Version",
-						Value:     fmt.Sprintf("%s vs %s", version, l.Value),
-					})
-				}
-			}
-		}
-
-		t := resume[s.Type]
-		t.nbOK += nbOK
-		t.nbWarn += nbWarn
-		t.nbAlerts += nbAlert
-		t.nbSrv++
-		resume[s.Type] = t
-
-		nbg.nbOK += nbOK
-		nbg.nbWarn += nbWarn
-		nbg.nbAlerts += nbAlert
-		nbg.nbSrv++
-	}
-
-	if versionOk {
-		linesGlobal = append(linesGlobal, sdk.MonitoringStatusLine{
-			Status:    sdk.MonitoringStatusOK,
-			Component: "Global/Version",
-			Value:     version,
-		})
-	}
-
-	linesGlobal = append(linesGlobal, sdk.MonitoringStatusLine{
-		Status:    api.computeGlobalStatusByNumbers(nbg),
-		Component: "Global/Status",
-		Value:     fmt.Sprintf("%d services", len(srvs)),
-	})
-
-	for stype, r := range resume {
-		linesGlobal = append(linesGlobal, sdk.MonitoringStatusLine{
-			Status:    api.computeGlobalStatusByNumbers(r),
-			Component: fmt.Sprintf("Global/%s", stype),
-			Value:     fmt.Sprintf("%d", r.nbSrv),
-		})
-	}
-
-	sort.Slice(linesGlobal, func(i, j int) bool {
-		return linesGlobal[i].Component < linesGlobal[j].Component
-	})
-
-	mStatus.Lines = append(linesGlobal, mStatus.Lines...)
-	return mStatus
-}
-
-func (api *API) computeGlobalStatusByNumbers(s computeGlobalNumbers) string {
-	r := sdk.MonitoringStatusOK
-	if s.nbAlerts > 0 {
-		r = sdk.MonitoringStatusAlert
-	} else if s.nbWarn > 0 {
-		r = sdk.MonitoringStatusWarn
-	} else if s.nbSrv < s.minInstance {
-		r = sdk.MonitoringStatusAlert
-	}
-	return r
 }
 
 func (api *API) smtpPingHandler() service.Handler {
